% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ser_from_univariate.R
\name{asymptotic_ser_from_univariate}
\alias{asymptotic_ser_from_univariate}
\title{Take a function for fitting univariate regression and make an SER function}
\usage{
asymptotic_ser_from_univariate(uni_fun)
}
\arguments{
\item{uni_fun}{a function for fitting univariate regression
must take arguments: x, y, o, prior_variance, estimate_intercept
additional arguments passed via ...
hint: it may be useful to pass null model likelihood in ...
must return: a list with items betahat = effect estimate, shat2 = square of standard errror, lr = likelihood ratio, intercept=intercept}
}
\description{
this is a utlity function for generating SER fit function compatible with \code{ibss_from_ser}
if you have a function for fitting a univariate regression this will create a function for fitting an SER
}
