% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_susie.R
\name{binsusie}
\alias{binsusie}
\title{Binomial SuSiE
Fit Binomial SuSiE via coordinate ascent variational inference}
\usage{
binsusie(
  X,
  y,
  N = rep(1, length(y)),
  Z = NULL,
  L = min(10, ncol(X)),
  prior_mean = 0,
  prior_variance = 1,
  prior_weights = NULL,
  intercept = TRUE,
  estimate_prior_variance = TRUE,
  s_init = NULL,
  coverage = 0.95,
  min_abs_corr = 0.5,
  max_iter = 100,
  tol = 0.001,
  verbose = FALSE,
  n_purity = 100
)
}
\arguments{
\item{X}{a n x p matrix of covariates}

\item{y}{an n vector of integer counts, bernoulli/binomial observations}

\item{N}{the number of binomial trials, defaults to 1, may be a scalar or vector of length n}

\item{Z}{fixed effect covaraites (including intercept). If null just a n x 1 matrix of ones}

\item{prior_mean}{the prior mean of each non-zero element of b. Either a scalar or vector of length L.}

\item{prior_variance}{the prior variance of each non-zero element of b. Either a scalar or vector of length L. If \code{estimate_prior_variance=TRUE} the value provides an initial estimate of prior variances}

\item{prior_weights}{prior probability of selecting each column of X, vector of length p summing to one, or an L x p matrix}

\item{s_init}{a logistic susie object to initialize with, NOTE if non-null, we ignore \code{prior_mean}, \code{prior_variance}, and \code{prior_weights}}

\item{returns}{a fit Binomial SuSiE model, which is compatable with summary functions from \code{susieR} package}
}
\description{
Binomial SuSiE
Fit Binomial SuSiE via coordinate ascent variational inference
}
