% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{binsusie_get_cs}
\alias{binsusie_get_cs}
\title{Get Credible Sets
Wraps susieR::susie_get_cs}
\usage{
binsusie_get_cs(
  fit,
  coverage = 0.95,
  min_abs_corr = 0.5,
  dedup = TRUE,
  squared = FALSE,
  check_symmetric = TRUE,
  n_purity = 100
)
}
\description{
Get Credible Sets
Wraps susieR::susie_get_cs
}
