% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise_initialization.R
\name{binsusie_init}
\alias{binsusie_init}
\title{Initialize binary SuSie}
\usage{
binsusie_init(
  X,
  y,
  N = rep(1, length(y)),
  Z = NULL,
  L = min(10, ncol(X)),
  scale = FALSE,
  center = TRUE,
  prior_mean = 0,
  prior_variance = 1,
  prior_weights = NULL
)
}
\description{
This is the main initialization function that takes arguments similar to \code{binsusie}, the main fitting/driver function
}
