% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise_initialization.R
\name{binsusie_init_xi_glm}
\alias{binsusie_init_xi_glm}
\title{initialize xi with the predictions from a glm using the tom \code{m} marginal enrichments}
\usage{
binsusie_init_xi_glm(fit, n_top = 10)
}
\description{
initialize xi with the predictions from a glm using the tom \code{m} marginal enrichments
}
