% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_susie.R
\name{binsusie_wrapup}
\alias{binsusie_wrapup}
\title{BinSuSiE Wrapup
Compute fit summaries (PIPs, CSs, etc) and
Organize binsusie fit so that it is compatable with \code{susieR} function}
\usage{
binsusie_wrapup(fit, prior_tol = 0)
}
\description{
BinSuSiE Wrapup
Compute fit summaries (PIPs, CSs, etc) and
Organize binsusie fit so that it is compatable with \code{susieR} function
}
