% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_helpers.R
\name{categorical_entropy}
\alias{categorical_entropy}
\title{Categorical Entropy
Compute entropy of Categorical(pi) = -E\link{log p}}
\usage{
categorical_entropy(pi)
}
\arguments{
\item{pi}{vector of probabilities that sum to 1}
}
\value{
Return entropy -Elog(x)
}
\description{
Categorical Entropy
Compute entropy of Categorical(pi) = -E\link{log p}
}
