% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{categorical_kl}
\alias{categorical_kl}
\title{Categorical KL
Compute KL(Categorical(alpha) || Categorical(pi))}
\usage{
categorical_kl(alpha, pi)
}
\arguments{
\item{alpha}{vector of probabilities that sum to 1}

\item{pi}{vector of probabilities that sum to 1}
}
\value{
Return KL-divergence KL(Categorical(alpha) || Categorical(pi))
}
\description{
Categorical KL
Compute KL(Categorical(alpha) || Categorical(pi))
}
