% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cEBMF.R
\name{cEBMF}
\alias{cEBMF}
\title{covariate moderated Empirical Bayes Matrix Factorization}
\usage{
cEBMF(
  Y,
  X_l,
  X_f,
  K = 1,
  type_noise = "constant",
  init_type = "udv_si",
  maxit = 100,
  tol = 0.001
)
}
\arguments{
\item{Y}{numerical matrix size NxP}

\item{X_l}{matrix of size NxJ containning covariates affecting the factors}

\item{X_f}{matrix of size PxT containning covariates affecting the factors}

\item{K}{numeric number of factors}

\item{type_noise}{specify which kind of noise structure is expected, currently three choices. Whether noise constant accross column ('column_wise'), constant 'constant' or constant across rown 'row_wise'}

\item{maxit}{maximum nuber of iterations}

\item{tol}{paramter for assessing convergence}
}
\value{
a cEBMF object
}
\description{
covariate moderated Empirical Bayes Matrix Factorization
}
