% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDR_wrappers.R
\name{cFDR}
\alias{cFDR}
\title{Main FDR function}
\usage{
cFDR(
  betahat,
  se,
  pvalue,
  zscore,
  X,
  maxiter = 100,
  tol = 0.001,
  max_class = 10,
  mult = 2,
  upper = FALSE
)
}
\description{
Main function gathering all the routines for different FDR estimation procedures
}
