% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_ser.R
\name{coef.binser}
\alias{coef.binser}
\title{Get Binomial SER coefficients
Extract regression coefficients from binomial SER fit}
\usage{
\method{coef}{binser}(fit, k, idx = NULL)
}
\arguments{
\item{fit}{Binomial SER object}
}
\value{
Return E\link[=\\beta]{\beta}
}
\description{
Get Binomial SER coefficients
Extract regression coefficients from binomial SER fit
}
