% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mococomo.R
\name{compute_assignment_loglikelihood.mococomo}
\alias{compute_assignment_loglikelihood.mococomo}
\title{Compute Expected Assignment Log Likelihood}
\usage{
compute_assignment_loglikelihood.mococomo(fit, normalize = TRUE)
}
\value{
N x K matrix each row has E\link[=p(z=k | \\beta, \\omega)]{p(z=k | \beta, \omega)} for k = 1,..., K
}
\description{
Compute Expected Assignment Log Likelihood
}
