% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_ser_wrapper.R
\name{compute_corrected_lbf}
\alias{compute_corrected_lbf}
\title{reestimate xi for each feature, this is the "corrected" SER}
\usage{
compute_corrected_lbf(X, y, fit)
}
\description{
reestimate xi for each feature, this is the "corrected" SER
}
