% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs.R
\name{compute_cs}
\alias{compute_cs}
\title{Compute credible sets}
\usage{
compute_cs(alpha, requested_coverage = 0.95)
}
\arguments{
\item{alpha}{a vector or matrix, where each row is a probability vector}

\item{requested_coverage}{the minimum probability mass the CS must satisfy}
}
\value{
a credible set object with the credible set and other summaries
if \code{alpha} is a vector, returns a list of credible set objects with names
\code{L1}, \code{L2}, ... etc.
}
\description{
Compute credible sets at a requested coverage level
}
