% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mococomo.R
\name{compute_data_loglikelihood}
\alias{compute_data_loglikelihood}
\title{Compute Data Log Likelihood
Compute p(betahat_i| z=k, se_i) i = 1..., n, k=1... K}
\usage{
compute_data_loglikelihood(fit)
}
\arguments{
\item{fit}{an MoCoCoMo fit object}
}
\value{
n x K matrix of  log likelihood of each data point for each component
}
\description{
Compute Data Log Likelihood
Compute p(betahat_i| z=k, se_i) i = 1..., n, k=1... K
}
