% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_ser.R
\name{compute_jj.binser}
\alias{compute_jj.binser}
\title{Compute E\link{p(z, w| b) - q(w)}, which is the same as
the bound on the logistic function proposed by Jaakkola and Jordan
NOTE: this only works when xi is updated!}
\usage{
\method{compute_jj}{binser}(fit, data)
}
\value{
a vector of length N with the lower bound for each data point
}
\description{
Compute E\link{p(z, w| b) - q(w)}, which is the same as
the bound on the logistic function proposed by Jaakkola and Jordan
NOTE: this only works when xi is updated!
}
