% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_susie.R
\name{compute_jj.binsusie}
\alias{compute_jj.binsusie}
\title{Compute E\link{p(z, w| b) - q(w)}, which is the same as
the bound on the logistic function proposed by Jaakkola and Jordan
NOTE: this only works when xi is updated!}
\usage{
\method{compute_jj}{binsusie}(fit, data)
}
\description{
Compute E\link{p(z, w| b) - q(w)}, which is the same as
the bound on the logistic function proposed by Jaakkola and Jordan
NOTE: this only works when xi is updated!
}
