% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_susie.R
\name{compute_jj2.binsusie}
\alias{compute_jj2.binsusie}
\title{Compute E\link{p(z, w| b) - q(w)}, which is the same as
the bound on the logistic function proposed by Jaakkola and Jordan}
\usage{
compute_jj2.binsusie(fit, data)
}
\description{
Compute E\link{p(z, w| b) - q(w)}, which is the same as
the bound on the logistic function proposed by Jaakkola and Jordan
}
