% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_ser.R
\name{compute_kl.binser}
\alias{compute_kl.binser}
\title{compute_kl.binser
Compute KL divergence between approximate posterior and prior}
\usage{
compute_kl.binser(fit)
}
\arguments{
\item{fit}{Binomial SER object}
}
\value{
Return KL(q(\beta) || p(\beta))
}
\description{
compute_kl.binser
Compute KL divergence between approximate posterior and prior
}
