% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_susie.R
\name{compute_kl.binsusie}
\alias{compute_kl.binsusie}
\title{Compute KL(q(\beta) || p(\beta)) for Sum of SERs
Note this does not include KL(q(\omega) || p(\omega))
since that gets computes as part of the JJ bound}
\usage{
compute_kl.binsusie(fit, k)
}
\description{
Compute KL(q(\beta) || p(\beta)) for Sum of SERs
Note this does not include KL(q(\omega) || p(\omega))
since that gets computes as part of the JJ bound
}
