% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_ser_wrapper.R
\name{compute_lbf_variable}
\alias{compute_lbf_variable}
\title{Compute log Bayes Factors for each feature
Takes a SER and give back the log-Bayes Factors for each feature}
\usage{
compute_lbf_variable(ser, l = 1)
}
\description{
Compute log Bayes Factors for each feature
Takes a SER and give back the log-Bayes Factors for each feature
}
