% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_ser.R
\name{compute_log_abf}
\alias{compute_log_abf}
\title{Compute log of the asymptotic Bayes factor (ABF) for the SER vs null model}
\usage{
compute_log_abf(betahat, s2, prior_variance)
}
\arguments{
\item{betahat}{vector of MLEs for the effect size}

\item{s2}{vector of standard errors}

\item{prior_variance}{variance of the effect size b ~ N(0, prior_variance)}
}
\value{
log ABF for each variable
}
\description{
Compute log of the asymptotic Bayes factor (ABF) for the SER vs null model
}
