% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ser_from_univariate.R
\name{compute_log_labf}
\alias{compute_log_labf}
\title{Compute log of the CORRECTED asymptotic Bayes factor (ABF)}
\usage{
compute_log_labf(betahat, s2, lr, prior_variance)
}
\arguments{
\item{betahat}{vector of MLEs for the effect size}

\item{s2}{vector of standard errors}

\item{prior_variance}{variance of the effect size b ~ N(0, prior_variance)}
}
\value{
log of the LABF for each variable
}
\description{
Compute log of the CORRECTED asymptotic Bayes factor (ABF)
}
