% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_quadrature.R
\name{compute_log_marginal}
\alias{compute_log_marginal}
\title{1d quadrature with a fixed intercept
Return the model evidence logp(y| x, b0, b_sigma)
which is the marginal likelihood for a logistic model
where we integrate over the effect b ~ N(0, b_sigma^2)}
\usage{
compute_log_marginal(x, y, b0, mu, var, prior_variance, n = 32, verbose = F)
}
\description{
1d quadrature with a fixed intercept
Return the model evidence logp(y| x, b0, b_sigma)
which is the marginal likelihood for a logistic model
where we integrate over the effect b ~ N(0, b_sigma^2)
}
