% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_quadrature.R
\name{compute_log_marginal_adaptive}
\alias{compute_log_marginal_adaptive}
\title{Adaptive quadrature
Double the number of quadrature points until change is small}
\usage{
compute_log_marginal_adaptive(
  x,
  y,
  b0,
  mu,
  var,
  prior_variance,
  n_init = 8,
  tol = 0.001,
  verbose = T
)
}
\description{
Adaptive quadrature
Double the number of quadrature points until change is small
}
