% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_susie.R
\name{compute_N.mnsusie}
\alias{compute_N.mnsusie}
\title{Conditional number of binomial trials}
\usage{
compute_N.mnsusie(Y)
}
\arguments{
\item{Y}{an N x K matrix of assignment probabilities}
}
\description{
Conditional number of binomial trials
}
