% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_ser.R
\name{compute_nu.binser}
\alias{compute_nu.binser}
\title{Compute a scaled posterior mean
TODO: add support for non-zero prior mean}
\usage{
compute_nu.binser(fit, k, idx = NULL, kidx = NULL, shift = 0)
}
\arguments{
\item{fit}{a SER object}
}
\description{
Compute a scaled posterior mean
TODO: add support for non-zero prior mean
}
