% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_generics.R
\name{compute_omega}
\alias{compute_omega}
\title{Compute E\link{w} where w are the PG random variables in the augmented model}
\usage{
compute_omega(fit, data)
}
\description{
Compute E\link{w} where w are the PG random variables in the augmented model
}
