% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_moococomo_obj.R
\name{compute_posterior_assignment}
\alias{compute_posterior_assignment}
\title{Compute Posterior Assignment Probabilities
For each data point return posterior assignment probabilities}
\usage{
compute_posterior_assignment(fit, log = TRUE)
}
\arguments{
\item{fit}{a MoCoCoMo fit object}
}
\value{
an n x K matrix of log posterior probabilities for each data point
}
\description{
Compute Posterior Assignment Probabilities
For each data point return posterior assignment probabilities
}
