% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_susie.R
\name{compute_prior_assignment}
\alias{compute_prior_assignment}
\title{Compute Log Prior Assignment Probabilities
For each data point return covariate-dependent prior mixture probabilities}
\usage{
compute_prior_assignment(fit)
}
\arguments{
\item{fit}{a multinomial SuSiE fit object (or MoCoCoMo)}
}
\value{
an n x K matrix of log prior probabilities for each data point
}
\description{
Compute Log Prior Assignment Probabilities
For each data point return covariate-dependent prior mixture probabilities
}
