% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mococomo.R
\name{compute_prior_assignment2}
\alias{compute_prior_assignment2}
\title{DONT USE, this ignores the PG agumentation and treats E\link{Xb_k} as the conditional log odds of sampling component k
and simply converts these quantitities to a vector of probabilities}
\usage{
compute_prior_assignment2(fit, log = TRUE)
}
\description{
DONT USE, this ignores the PG agumentation and treats E\link{Xb_k} as the conditional log odds of sampling component k
and simply converts these quantitities to a vector of probabilities
}
