% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_generics.R
\name{compute_tau}
\alias{compute_tau}
\title{Compute partial posterior variance
use this immediately after updating xi}
\usage{
compute_tau(fit, data)
}
\description{
Compute partial posterior variance
use this immediately after updating xi
}
