% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_ser.R
\name{compute_Xb.binser}
\alias{compute_Xb.binser}
\title{compute_Xb.binser
Compute expected predicted log odds}
\usage{
\method{compute_Xb}{binser}(fit, data)
}
\arguments{
\item{fit}{Binomial SER object}
}
\value{
Return E\link{Xb}
}
\description{
compute_Xb.binser
Compute expected predicted log odds
}
