% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_ser.R
\name{compute_Xb2.binser}
\alias{compute_Xb2.binser}
\title{compute_Xb2.binser
Compute second moments of predicted log odds}
\usage{
\method{compute_Xb2}{binser}(fit, data)
}
\arguments{
\item{fit}{Binomial SER object}
}
\value{
Return E\link{Xb}
}
\description{
compute_Xb2.binser
Compute second moments of predicted log odds
}
