% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_distributions.R
\name{convolved_logpdf.normal}
\alias{convolved_logpdf.normal}
\title{Convolved log density}
\usage{
\method{convolved_logpdf}{normal}(dist, betahat, se)
}
\arguments{
\item{dist}{is a list with parameters for the normal component}

\item{betahat}{vector of oberveations}

\item{se}{vector of standard errors of observations}
}
\value{
A vector of log densities log p(\hat\beta | se) = log N \hat\beta ; \mu, sigma^2 + se^2)
}
\description{
Convolved log density
}
