% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_distributions.R
\name{convolved_logpdf}
\alias{convolved_logpdf}
\title{Convolved logpdf}
\usage{
convolved_logpdf(x, ...)
}
\arguments{
\item{betahat}{observations}

\item{se}{standard errors of observations}
}
\value{
the convolved log density log p(betahat | se) = log \int N(betahat | beta, se) p(beta) dbeta
}
\description{
Generic function for \code{component_distribution} objects, computes the loglikelihood of an observation from
the component distribution, corrupted by gaussian measurement error
}
