% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_ser_wrapper.R
\name{correct_bin_ser}
\alias{correct_bin_ser}
\title{Correct a binomial SER fit by recomputing the lbfs}
\usage{
correct_bin_ser(X, y, o, fit)
}
\description{
Correct a binomial SER fit by recomputing the lbfs
}
