% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_function.R, R/multinomial_susie.R
\name{.compute_prior_assignment}
\alias{.compute_prior_assignment}
\title{Compute Log Prior Assignment Probabilities
For each data point return covariate-dependent prior mixture probabilities}
\usage{
.compute_prior_assignment(xb)

.compute_prior_assignment(xb)
}
\arguments{
\item{xb}{a K-1 vector of predictions}

\item{fit}{a MoCoCoMo fit object}
}
\value{
an n x K matrix of log prior probabilities for each data point

an n x K matrix of log prior probabilities for each data point
}
\description{
Compute Log Prior Assignment Probabilities
For each data point return covariate-dependent prior mixture probabilities

Compute Log Prior Assignment Probabilities
For each data point return covariate-dependent prior mixture probabilities
}
