% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_moococomo_obj.R
\name{.expected_trials}
\alias{.expected_trials}
\title{Expected Binomial Trials
Compute expected  number of binomial trials
MoCoCoMo fits K-1 binomial regressions, this is the E-step for N}
\usage{
.expected_trials(Y)
}
\arguments{
\item{Y}{an N x K matrix of assignment probabilities}
}
\description{
Expected Binomial Trials
Compute expected  number of binomial trials
MoCoCoMo fits K-1 binomial regressions, this is the E-step for N
}
