% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise_initialization.R
\name{.forward_init}
\alias{.forward_init}
\title{initialize SER}
\usage{
.forward_init(
  data,
  L = 5,
  init_iter = 10,
  prior_mean = 0,
  prior_variance = 1,
  prior_weights = NULL,
  kidx = NULL
)
}
\description{
initialize SER
}
