% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss_hermite.R
\name{.gauss_hermite}
\alias{.gauss_hermite}
\title{perform 2d gauss hermite quadrature for a given likelihood function
B is a list with nodes and weights}
\usage{
.gauss_hermite(llfun, B)
}
\description{
perform 2d gauss hermite quadrature for a given likelihood function
B is a list with nodes and weights
}
