% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss_hermite_2d.R
\name{.gauss_hermite2d}
\alias{.gauss_hermite2d}
\title{perform 2d gauss hermite quadrature for a given likelihood function
B is a list with nodes and weights}
\usage{
.gauss_hermite2d(llfun, B)
}
\description{
perform 2d gauss hermite quadrature for a given likelihood function
B is a list with nodes and weights
}
