% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mococomo.R
\name{.predict2logpi}
\alias{.predict2logpi}
\title{Predict to log pi
Computes E\link[=log p(z | \\beta, \\omega)]{log p(z | \beta, \omega)} from E\link{Xb_1}, ... E\link{Xb_{K-1}}
Where expectations are taken under variational approximation}
\usage{
.predict2logpi(xb)
}
\arguments{
\item{xb}{vector, length K-1 of conditional log odds for first K-1 components}
}
\value{
a vector of length K of (expected) log prior assignment probabilities
}
\description{
Predict to log pi
Computes E\link[=log p(z | \\beta, \\omega)]{log p(z | \beta, \omega)} from E\link{Xb_1}, ... E\link{Xb_{K-1}}
Where expectations are taken under variational approximation
}
