% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_ser.R
\name{.update_b_ser}
\alias{.update_b_ser}
\title{update SER variational parameters}
\usage{
.update_b_ser(nu, tau, pi)
}
\arguments{
\item{pi}{prior probability vector (length p)}
}
\description{
update SER variational parameters
}
