% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_ser.R
\name{fit.binser}
\alias{fit.binser}
\title{Fit the binomial single effect regression}
\usage{
fit.binser(
  data = NULL,
  fit = NULL,
  maxiter = 10,
  tol = 0.001,
  fit_intercept = TRUE,
  fit_prior_variance = TRUE,
  shift = NULL
)
}
\description{
Fit the binomial single effect regression
}
