% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_susie.R
\name{fit.binsusie}
\alias{fit.binsusie}
\title{Fit the binomial single effect regression}
\usage{
fit.binsusie(
  data,
  maxiter = 10,
  tol = 0.001,
  fit_intercept = TRUE,
  fit_prior_variance = TRUE,
  kidx = NULL
)
}
\description{
Fit the binomial single effect regression
}
