% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_susie.R
\name{fit.mnsusie}
\alias{fit.mnsusie}
\title{Fit the binomial single effect regression}
\usage{
fit.mnsusie(
  data,
  L = 5,
  maxiter = 10,
  tol = 0.001,
  fit_intercept = TRUE,
  fit_prior_variance = TRUE
)
}
\description{
Fit the binomial single effect regression
}
