% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_bayes.R
\name{fit_bayes_ser}
\alias{fit_bayes_ser}
\title{Fit logistic SER via quadrature}
\usage{
fit_bayes_ser(
  X,
  y,
  o = NULL,
  prior_variance = 1,
  estimate_intercept = T,
  prior_weights = NULL,
  n = 2^10,
  optimize_b0 = F
)
}
\description{
Estimate a fixed intercept using UVB approximation
and then numerically integrate over effect variable with normal prior
}
