% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_ser_wrapper.R
\name{fit_bin_ser_corrected}
\alias{fit_bin_ser_corrected}
\title{Fit a logistic single effect regression with our variational approximation
Returns a list with posterior summary}
\usage{
fit_bin_ser_corrected(
  X,
  y,
  o = NULL,
  prior_variance = 1,
  estimate_intercept = T,
  prior_weights = NULL
)
}
\description{
Fit a logistic single effect regression with our variational approximation
Returns a list with posterior summary
}
