% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_ser.R
\name{fit_glm_ser2}
\alias{fit_glm_ser2}
\title{compute SER using GLM, use Laplace approximation to BF rather than ABF}
\usage{
fit_glm_ser2(
  X,
  y,
  o = NULL,
  prior_variance = 1,
  estimate_intercept = T,
  prior_weights = NULL,
  family = "binomial"
)
}
\description{
compute SER using GLM, use Laplace approximation to BF rather than ABF
}
