% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_quadrature.R
\name{fit_quad_1d_fixed_int}
\alias{fit_quad_1d_fixed_int}
\title{1d quadrature with a fixed intercept
Return the model evidence logp(y| x, b0, b_sigma)
which is the marginal likelihood for a logistic model
where we integrate over the effect b ~ N(0, b_sigma^2)}
\usage{
fit_quad_1d_fixed_int(x, y, b0, b_sigma, n = 2^10)
}
\description{
1d quadrature with a fixed intercept
Return the model evidence logp(y| x, b0, b_sigma)
which is the marginal likelihood for a logistic model
where we integrate over the effect b ~ N(0, b_sigma^2)
}
