% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_quadrature.R
\name{fit_quad_ser}
\alias{fit_quad_ser}
\title{Fit quad SER}
\usage{
fit_quad_ser(
  X,
  y,
  o = NULL,
  prior_variance = 1,
  intercept = T,
  prior_weights = NULL,
  glm_ser = NULL,
  n = 16,
  verbose = F,
  ...
)
}
\arguments{
\item{X}{n x p matrix of covariates}

\item{y}{n vector of responses}

\item{o}{n vector of fixed offsets}

\item{glm_ser}{provide a fit glm_ser object to avoid refitting}

\item{n}{number of quadrature points}

\item{verbose}{print messages during fitting}

\item{glm_ser_args}{arguments to be passed to \code{fit_glm_ser2}}
}
\description{
posterior mean and variance are approximated using an asymptotic normal approximation
the same as in \code{fit_glm_ser2}
the BFs are computed via quadrature, with a quadrature rule
centered on the approximate MAP
}
