% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_random_effect_vb.R
\name{fit_univariate_vb_re}
\alias{fit_univariate_vb_re}
\title{Quick implimentation of univariate VB logistic regression
The intercept \code{delta} is treated as a parameter to be optimized,
Normal prior on the effect N(0, 1/tau0)}
\usage{
fit_univariate_vb_re(
  x,
  y,
  o = NULL,
  delta.init = logodds(mean(y) + 1e-10),
  tau0 = 1,
  estimate_intercept = T,
  maxit = 50,
  tol = 0.001
)
}
\description{
Quick implimentation of univariate VB logistic regression
The intercept \code{delta} is treated as a parameter to be optimized,
Normal prior on the effect N(0, 1/tau0)
}
