% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_vb.R
\name{fit_uvb_ser}
\alias{fit_uvb_ser}
\title{Fit a logistic single effect regression model using univariate VB approximation}
\usage{
fit_uvb_ser(
  X,
  y,
  o = NULL,
  o2 = NULL,
  prior_variance = 1,
  intercept.init = logodds(mean(y) + 1e-10),
  estimate_intercept = T,
  prior_weights = NULL
)
}
\description{
Fit a logistic single effect regression model using univariate VB approximation
}
