% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_random_effect_vb.R
\name{fit_uvb_ser_re}
\alias{fit_uvb_ser_re}
\title{Fit a logistic single effect regression model using univariate VB approximation}
\usage{
fit_uvb_ser_re(
  X,
  y,
  o = NULL,
  prior_variance = 1,
  intercept.init = logodds(mean(y) + 1e-10),
  estimate_intercept = T,
  estimate_prior_variance = F,
  prior_weights = NULL
)
}
\description{
Fit a logistic single effect regression model using univariate VB approximation
}
