% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss_hermite.R
\name{gauss_hermite}
\alias{gauss_hermite}
\title{perform 2d gauss hermite quadrature for a given likelihood function
nodes are determined using a local quadratic approximation N(mu, H)}
\usage{
gauss_hermite(llfun, tau, mu, m, Z = NULL)
}
\description{
perform 2d gauss hermite quadrature for a given likelihood function
nodes are determined using a local quadratic approximation N(mu, H)
}
