% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss_hermite_2d.R
\name{gauss_hermite2d}
\alias{gauss_hermite2d}
\title{perform 2d gauss hermite quadrature for a given likelihood function
nodes are determined using a local quadratic approximation N(mu, H)}
\usage{
gauss_hermite2d(llfun, H, mu, m1, m2, type = "svd", Z = NULL)
}
\description{
perform 2d gauss hermite quadrature for a given likelihood function
nodes are determined using a local quadratic approximation N(mu, H)
}
