% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_quadrature.R
\name{gauss_quad2}
\alias{gauss_quad2}
\title{Gaussian quadrature}
\usage{
gauss_quad2(log_f, mu, sd, n = 32)
}
\arguments{
\item{mu}{mean of gaussian quadrature}

\item{sd}{sd of gaussian quadrature}
}
\description{
approximate log(\int f(x)dx) by quadrature on \int f(x)/p(x) p(x)
where p(x) is a normal density, we approximate this integral
with a Gauss-Hermite quadrature rule via \code{statmod::gauss.quad.prob}
done on a log scale for numerical stability (f is positive)
the goal is to choose p(x) such that f(x)/p(x) is not too variable near the mode
in practice we use the Laplace approximation of the posterior for p(x)
}
