% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibss_ser.R
\name{generalized_ibss}
\alias{generalized_ibss}
\title{Generalized IBSS}
\usage{
generalized_ibss(X, y, L = 10, tol = 1e-08, maxit = 100, init = NULL, ...)
}
\arguments{
\item{X}{design matrix}

\item{y}{response}

\item{L}{number of single effects}

\item{laplace}{boolean to use Laplace approximation to BF rather than ABF--
we recommend keeping set to default \code{TRUE}}

\item{estimate_prior_variance}{boolean to estimate prior variance}

\item{family}{family for glm}
}
\description{
approximate GLM SuSiE using generalized IBSS heuristic
by default this is for logistic regression but you can specify
what GLM to use via "family" argument
dots get passed to \code{ibss_from_ser} see documentation for options
}
