% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibss_ser.R
\name{generalized_ibss_quad}
\alias{generalized_ibss_quad}
\title{Generalized IBSS Quadrature}
\usage{
generalized_ibss_quad(
  X,
  y,
  L = 10,
  n = 32,
  verbose = F,
  estimate_prior_variance = T,
  min_prior_variance = 0,
  family = "binomial",
  ...
)
}
\arguments{
\item{X}{design matrix}

\item{y}{response}

\item{L}{number of single effects}

\item{estimate_prior_variance}{boolean to estimate prior variance}

\item{family}{family for glm}

\item{laplace}{boolean to use Laplace approximation to BF rather than ABF--
we recommend keeping set to default \code{TRUE}}
}
\description{
approximate GLM SuSiE using generalized IBSS heuristic
by default this is for logistic regression but you can specify
what GLM to use via "family" argument
dots get passed to \code{ibss_from_ser} see documentation for options
}
