% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_all_coverage}
\alias{get_all_coverage}
\title{Check if set of indices \code{idx} in a list of credible sets}
\usage{
get_all_coverage(css, idx)
}
\description{
Check if set of indices \code{idx} in a list of credible sets
}
