% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_susie.R
\name{get_xi.cbsusie}
\alias{get_xi.cbsusie}
\title{Get N x K-1 matrix of PG variational parameters
TODO: deprecate this when we call update functions directly rather than
passing to logistic susie subroutine}
\usage{
get_xi.cbsusie(fit)
}
\description{
Get N x K-1 matrix of PG variational parameters
TODO: deprecate this when we call update functions directly rather than
passing to logistic susie subroutine
}
