% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_moococomo_obj.R
\name{get_xi.mococomo}
\alias{get_xi.mococomo}
\title{Get N x K-1 matrix of PG variational parameters
TODO: deprecate this when we call update functions directly rather than
passing to logistic susie subroutine}
\usage{
get_xi.mococomo(fit)
}
\description{
Get N x K-1 matrix of PG variational parameters
TODO: deprecate this when we call update functions directly rather than
passing to logistic susie subroutine
}
