% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbnm_susie.R
\name{get_xi.sbnmsusie}
\alias{get_xi.sbnmsusie}
\title{Get N x K-1 matrix of PG variational parameters
TODO: deprecate this when we call update functions directly rather than
passing to logistic susie subroutine}
\usage{
get_xi.sbnmsusie(fit)
}
\description{
Get N x K-1 matrix of PG variational parameters
TODO: deprecate this when we call update functions directly rather than
passing to logistic susie subroutine
}
