% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise_initialization.R
\name{hard_select_forward}
\alias{hard_select_forward}
\title{update alpha, mu, and var}
\usage{
hard_select_forward(
  fit,
  fit_intercept = TRUE,
  fit_prior_variance = TRUE,
  update_idx = NULL
)
}
\description{
update alpha, mu, and var
}
