% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_ser.R
\name{init.binser}
\alias{init.binser}
\title{initialize SER}
\usage{
init.binser(data)
}
\arguments{
\item{data}{a list containing X, Z, y}

\item{prior_mean}{prior effect mean}

\item{simga0}{prior effect standard deviation}
}
\description{
initialize SER
}
