% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_susie.R
\name{init.binsusie}
\alias{init.binsusie}
\title{initialize SER}
\usage{
init.binsusie(
  data,
  L = 5,
  prior_mean = 0,
  prior_variance = 1,
  prior_weights = NULL,
  kidx = NULL,
  from_init_moco = FALSE
)
}
\description{
initialize SER
}
