% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_twococomo_obj.R
\name{init.twococomo}
\alias{init.twococomo}
\title{Initialize Two CoCoMo}
\usage{
init.twococomo(
  data,
  L = 5,
  f1_dist = "normal",
  f1_params = list(),
  prior_mean = 0,
  prior_variance = 1,
  prior_weights = NULL
)
}
\arguments{
\item{data}{list of data for twococomo}

\item{L}{number of single effects}

\item{f1_dist}{the name of \code{component_distribution} for the non-null component}

\item{f1_params}{list of parameters to initialize f1, must be a subset of those accepted by the constructor for f1}

\item{prior_mean}{the prior mean of each non-zero element of b. Either a scalar or vector of length L.}

\item{prior_variance}{the prior variance of each non-zero element of b. Either a scalar or vector of length L. If \code{estimate_prior_variance=TRUE} the value provides an initial estimate of prior variances}

\item{prior_weights}{prior probability of selecting each column of X, vector of length p summing to one, or an L x p matrix}

\item{returns}{am initialized twococomo object}
}
\description{
Initialize Two CoCoMo
}
