% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_cEBMF_obj.R
\name{init_cEBMF}
\alias{init_cEBMF}
\title{Initialize cEBMF object}
\usage{
init_cEBMF(Y, X_l, X_f, K = 1, type_noise = "constant", init_type = "udv_si")
}
\arguments{
\item{Y}{numerical matrix size NxP}

\item{X_l}{matrix of size NxJ containning covariates affecting the factors}

\item{X_f}{matrix of size PxT containning covariates affecting the factors}

\item{K}{numeric number of factors
@param type_noise specify which kind of noise structure is expected, currently three choices. Whether noise constant accross column ('column_wise'), constant 'constant' or constant across rown 'row_wise'
@param init_type specify initialisation method for loading and factor, methods includes: "udv", "udv_si_svd", "si" (softimpute), "ones" and "rand_udv"}
}
\value{
a cEBMF object
}
\description{
Initialize cEBMF object
}
