% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso_init.R
\name{initialize_q_with_lasso}
\alias{initialize_q_with_lasso}
\title{Initialize q with lasso}
\usage{
initialize_q_with_lasso(X, y, L = "auto", family = "binomial")
}
\description{
Providies initialization for q using lasso
}
