% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_ser.R
\name{jj_bound.binser}
\alias{jj_bound.binser}
\title{Compute E\link{p(z, w| b) - q(w)}, which is the same as
the bound on the logistic function proposed by Jaakkola and Jordan
NOTE: this only works when xi is updated!}
\usage{
jj_bound.binser(fit, kidx = NULL)
}
\value{
a vector of length N with the lower bound for each data point
}
\description{
Compute E\link{p(z, w| b) - q(w)}, which is the same as
the bound on the logistic function proposed by Jaakkola and Jordan
NOTE: this only works when xi is updated!
}
