% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_susie.R
\name{jj_bound.binsusie}
\alias{jj_bound.binsusie}
\title{Compute E\link{p(z, w| b) - q(w)}, which is the same as
the bound on the logistic function proposed by Jaakkola and Jordan
NOTE: this only works when xi is updated!}
\usage{
jj_bound.binsusie(fit)
}
\description{
Compute E\link{p(z, w| b) - q(w)}, which is the same as
the bound on the logistic function proposed by Jaakkola and Jordan
NOTE: this only works when xi is updated!
}
