% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_ser.R
\name{jj_bound.logistic}
\alias{jj_bound.logistic}
\title{More explicit, only works for Bernoulli observations}
\usage{
jj_bound.logistic(fit, kidx = NULL)
}
\description{
More explicit, only works for Bernoulli observations
}
