% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_susie.R
\name{jj_bound2.binsusie}
\alias{jj_bound2.binsusie}
\title{Compute E\link{p(z, w| b) - q(w)}, which is the same as
the bound on the logistic function proposed by Jaakkola and Jordan}
\usage{
jj_bound2.binsusie(fit)
}
\description{
Compute E\link{p(z, w| b) - q(w)}, which is the same as
the bound on the logistic function proposed by Jaakkola and Jordan
}
