% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_correction_alpha.R
\name{linear_correction_alpha3}
\alias{linear_correction_alpha3}
\title{We've seen that the variational approximation under-covers
Here we implement an experimental correction based on the observed
linear relationship between the uni variate vb log BFs
which provide a tight bound to the true BFs and can be compared against each-other
Our strategy is to compute the exact BFs for the top m features and compute a linear correction term}
\usage{
linear_correction_alpha3(fit, m = 5, idx = NULL)
}
\arguments{
\item{fit}{a binsusie object}

\item{m}{the number of top features to estimate the correction factor in}

\item{idx}{which single effect to update, 1, ..., L}
}
\description{
We've seen that the variational approximation under-covers
Here we implement an experimental correction based on the observed
linear relationship between the uni variate vb log BFs
which provide a tight bound to the true BFs and can be compared against each-other
Our strategy is to compute the exact BFs for the top m features and compute a linear correction term
}
