% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_ridge.R
\name{logistic_bayes}
\alias{logistic_bayes}
\title{fit ridge regression and then numerically compute: mean, variance, normalizing constant}
\usage{
logistic_bayes(x, y, o = NULL, prior_variance = 1, eps = 0, width = 5)
}
\description{
fit ridge regression and then numerically compute: mean, variance, normalizing constant
}
