% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_ridge.R
\name{logistic_ser}
\alias{logistic_ser}
\title{slow but (hopefully) accurate logistic SER with normal prior}
\usage{
logistic_ser(X, y, o = NULL, prior_variance = 1)
}
\description{
slow but (hopefully) accurate logistic SER with normal prior
}
