% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss_hermite.R
\name{logistic_ser_hermite}
\alias{logistic_ser_hermite}
\title{Logistic regression with a gaussian prior on the effect
intercept is fixed to the MAP estimate (note: intercept is not regularized, but the effect is)
computed via gauss hermite quadrature
equivalent to Laplace approximation when m=1}
\usage{
logistic_ser_hermite(X, y, o = NULL, prior_variance = 1, m = 5)
}
\arguments{
\item{X}{design matrix}

\item{y}{binary response}

\item{o}{offset}

\item{prior_variance}{prior variance of the effect}

\item{m}{number of quadrature points}
}
\description{
Logistic regression with a gaussian prior on the effect
intercept is fixed to the MAP estimate (note: intercept is not regularized, but the effect is)
computed via gauss hermite quadrature
equivalent to Laplace approximation when m=1
}
