% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss_hermite_2d.R
\name{logistic_ser_hermite2d}
\alias{logistic_ser_hermite2d}
\title{Bayesian logistic regression with independent gaussian priors on the intercept and effect
Numerical integration via Gauss-Hermite quadrature}
\usage{
logistic_ser_hermite2d(
  X,
  y,
  o = NULL,
  prior_variance = 1,
  prior_variance_b0 = 1e+05,
  m1 = 5,
  m2 = 5,
  type = "svd"
)
}
\arguments{
\item{X}{matrix of covariates}

\item{y}{binary response}

\item{o}{offset}

\item{prior_variance}{variance of the prior}

\item{prior_variance_b0}{variance of the prior for the intercept}

\item{m1}{number of quadrature points for the intercept}

\item{m2}{number of quadrature points for the effect}

\item{type}{type of transformation for the quadrature points, one of 'svd', 'chol', 'diag'}
}
\description{
Bayesian logistic regression with independent gaussian priors on the intercept and effect
Numerical integration via Gauss-Hermite quadrature
}
