% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_ridge.R
\name{make_laplace_approx}
\alias{make_laplace_approx}
\title{make quadratic approx fun (as a function of b)}
\usage{
make_laplace_approx(log_joint, ridgefit)
}
\description{
make quadratic approx fun (as a function of b)
}
