% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_ridge.R
\name{make_log_joint2d}
\alias{make_log_joint2d}
\title{make log likelihood function (as a function of b0, b1)}
\usage{
make_log_joint2d(x, y, o, prior_variance)
}
\description{
make log likelihood function (as a function of b0, b1)
}
