% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_ridge.R
\name{make_log_joint_ridge}
\alias{make_log_joint_ridge}
\title{make log likelihood function (as a function of b)}
\usage{
make_log_joint_ridge(x, y, o, ridgefit)
}
\description{
make log likelihood function (as a function of b)
}
