% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_random_effect_vb.R
\name{make_post_re}
\alias{make_post_re}
\title{extract posterior summaries from list of uvb_re fits}
\usage{
make_post_re(fits)
}
\description{
extract posterior summaries from list of uvb_re fits
}
