% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{mnsusie_prep_data}
\alias{mnsusie_prep_data}
\title{Prepare data for multinomial regression}
\usage{
mnsusie_prep_data(
  X,
  Y,
  Z = NULL,
  shift = 0,
  shift_var = 0,
  center = TRUE,
  scale = FALSE
)
}
\arguments{
\item{X}{matrix of covoariates}

\item{Z}{matrix of fixed-effect covariates (including intercept)}

\item{center}{boolean to center \code{X}}

\item{scale}{boolean to scale \code{X} to unit variance}

\item{y}{matrix of multinomial observations (each row an observation)}

\item{N}{vector giving number of trials for each multinomial observation}
}
\description{
Prepare data for multinomial regression
}
