% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{normal_kl}
\alias{normal_kl}
\title{Normal KL
Compute KL(N(mu, var) || N(mu0, var0))}
\usage{
normal_kl(mu, var, mu0 = 0, var0 = 1)
}
\arguments{
\item{mu}{first mean parameter of}

\item{var}{fist variance parameter}

\item{mu0}{second mean parameter}

\item{var0}{second variance parameter}
}
\value{
KL(N(mu, var) || N(mu0, var0))
}
\description{
Normal KL
Compute KL(N(mu, var) || N(mu0, var0))
}
