% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_ser.R
\name{optimize_prior_variance}
\alias{optimize_prior_variance}
\title{Optimize prior variance}
\usage{
optimize_prior_variance(betahat, s2, lr, pi, laplace = T)
}
\arguments{
\item{betahat}{vector of MLEs for the effect size}

\item{s2}{vector of standard errors}

\item{pi}{prior over non-zero effects}
}
\value{
the optimimum prior variance
}
\description{
Optimize the asymptotic BF for the SER w.r.t the prior variance
}
