% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_cEBMF_obj.R
\name{out_prep.cEBMF}
\alias{out_prep.cEBMF}
\title{prepare output of cEMBF function}
\usage{
out_prep.cEBMF(cEBMF.obj)
}
\arguments{
\item{cEBMF.obj}{a cEBMF object}
}
\value{
a cEBMF object
}
\description{
prepare output of cEMBF function
}
