% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pg_kl}
\alias{pg_kl}
\title{Polya-Gamma KL
Compute KL(PG(b, c) || N(b, 0))}
\usage{
pg_kl(b, c)
}
\arguments{
\item{b}{first mean parameter of}

\item{c}{fist variance parameter}
}
\value{
Return KL-divergence KL(PG(b, c) || N(b, 0))
}
\description{
Polya-Gamma KL
Compute KL(PG(b, c) || N(b, 0))
}
