% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_function.R
\name{pg_mean}
\alias{pg_mean}
\title{polya Gamma Mean
Compute the mean of a Polya Gamma RV w ~ PG(b, c)
@param b a PG parameter
@param c a PG parameter
@return E\link{w} the mean of w ~ PG(b, c)}
\usage{
pg_mean(b, c)
}
\description{
polya Gamma Mean
Compute the mean of a Polya Gamma RV w ~ PG(b, c)
@param b a PG parameter
@param c a PG parameter
@return E\link{w} the mean of w ~ PG(b, c)
}
