% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_moococomo_obj.R
\name{post_mean_sd.mococomo}
\alias{post_mean_sd.mococomo}
\title{Compute individual posterior mean and sd under mococomo model}
\usage{
post_mean_sd.mococomo(fit)
}
\arguments{
\item{fit}{a mococomo object}
}
\description{
Compute individual posterior mean and sd under mococomo model

# TODO: allow using new observation from another data set (e.g. testing)
}
