% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_susie.R
\name{predict_log_assignment_cbsusie}
\alias{predict_log_assignment_cbsusie}
\title{Compute E\link{log p(z | X, B)}}
\usage{
predict_log_assignment_cbsusie(fit, data)
}
\description{
Compute E\link{log p(z | X, B)}
}
