% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_ser.R
\name{prep_data_for_fastglm}
\alias{prep_data_for_fastglm}
\title{Prepare data for fast glm}
\usage{
prep_data_for_fastglm(x, y, o, intercept = T, augment = F)
}
\arguments{
\item{x}{the variable of interest}

\item{y}{the response vector}

\item{o}{fixed offset, length(o) == length(y)}

\item{intercept}{boolean to include column of 1s in design matrix}

\item{augment}{boolean to augment data to ensure existence of MLE}
}
\value{
a list containing design \code{X}, reponse \code{y}, and offset \code{o}
}
\description{
Helper function for preparing the design matrix, response vector, and offset
}
