% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_susie.R
\name{prune_model}
\alias{prune_model}
\title{Prune Model
Remove irrelevant features from logistic SuSiE fit}
\usage{
prune_model(
  fit,
  k,
  check_null_threshold,
  fit_intercept = T,
  fit_prior_variance = F,
  tol = 0.001
)
}
\arguments{
\item{fit}{a fit logistic Susie model}

\item{check_null_threshold}{if ELBO_null + check_null_threshold > ELBO drop the single effect}

\item{fit_intercept}{boolean to re-estimate intercept in simplified model, defaul TRUE}

\item{fit_prior_variance}{boolean to re-estimate prior variance in simplified model, default FALSE}

\item{tol}{tolerance to declare convergence when fitting simplified model}
}
\description{
Prune Model
Remove irrelevant features from logistic SuSiE fit
}
