% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_susie.R
\name{prune_model_idx}
\alias{prune_model_idx}
\title{remove a single effect and refit model
This function zeros out a single effect and refits the model
Used as a subroutine for \code{prune_model} which removes irrelevant features}
\usage{
prune_model_idx(
  fit,
  k,
  idx,
  fit_intercept = T,
  fit_prior_variance = F,
  tol = 0.001,
  maxiter = 10
)
}
\description{
remove a single effect and refit model
This function zeros out a single effect and refits the model
Used as a subroutine for \code{prune_model} which removes irrelevant features
}
