% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibss_ser.R
\name{ser_from_univariate}
\alias{ser_from_univariate}
\title{Take a function for fitting univariate regression and}
\usage{
ser_from_univariate(uni_fun)
}
\arguments{
\item{uni_fun}{a function for fitting univariate regression
must take arguments: x, y, o, prior_variance, estimate_intercept
additional arguments passed via ...
hint: it may be useful to pass null model likelihood in ...
must return: a list with items lbf = log bayes factor, mu = posterior mean, var = posterior variance, intercept=intercept
note that posterior variance doesn't figure in computations so you can set it to a dummy value (e.g. 0)}
}
\description{
this is a utlity function for generating SER fit function compatible with \code{ibss_from_ser}
if you have a function for fitting a univariate regression this will create a function for fitting an SER
}
