% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{set_data_mococomo}
\alias{set_data_mococomo}
\title{Preparing data for mococomo fit}
\usage{
set_data_mococomo(betahat, se, p, X, ...)
}
\arguments{
\item{betahat}{numeric, vector of regression coefficients  list of containing the following element see}

\item{se}{numeric, corresponding standard error}

\item{p}{numeric observed p-values}

\item{maxiter}{numeric, maximum numerous of iteration set to 100 by defaults}

\item{tol}{tolerance in term of change in ELBO value for stopping criterion}
}
\description{
Preparing data for mococomo fit
}
\details{
Preparing data for mococomo fit for two type of input p-values or estimated regression coefficients
currently only support one type of entry, whether betahat or p-values
}
