% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sigmoid}
\alias{sigmoid}
\title{Sigmoid
sigmoid function coverts log-odds to probability}
\usage{
sigmoid(x)
}
\arguments{
\item{x}{Log odds}
}
\value{
Returns the probability
}
\description{
Sigmoid
sigmoid function coverts log-odds to probability
}
