% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{sim_mixture_beta}
\alias{sim_mixture_beta}
\title{Simulate mixture of beta distribution}
\usage{
sim_mixture_beta(
  n = 1000,
  par = list(alpha = c(1, 1), beta = c(1, 10)),
  prob = c(0.9, 0.1)
)
}
\arguments{
\item{n}{numeric sample size}

\item{par}{list of the component parameter}

\item{prob}{mixture proportion}
}
\description{
Simulate mixture of beta distribution
}
