% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{sim_mococomo_beta}
\alias{sim_mococomo_beta}
\title{Simulate data under mococomo model with beta distribution}
\usage{
sim_mococomo_beta(n = 1000, p = 50, L = 3, N = 1, beta)
}
\description{
Simulate data under mococomo model with beta distribution
}
\details{
Simulate data under mococomo model with beta distribution
}
