% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{sim_ser_with_covariates}
\alias{sim_ser_with_covariates}
\title{simulate SER with 3 covariates}
\usage{
sim_ser_with_covariates(n = 1000, p = 50, N = 1, idx = 1)
}
\description{
simulate SER with 3 covariates
}
