% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{sim_twococomo}
\alias{sim_twococomo}
\title{Simulate data under the twococomo SuSiE model}
\usage{
sim_twococomo(n = 1000, p = 50, L = 3, N = 1, beta, alpha = 0)
}
\arguments{
\item{n}{numeric sample size}

\item{p}{numeric number of observed covariates}

\item{L}{numeric number of causal covariates (should be lower than p)}

\item{N}{numeric number of draw in the bionmial considered (set to 1 as default)}

\item{beta}{numeric if of length 1 assume that ervery effect have same beta coefficietns
otherwise should be  of length L, if missing beta <- seq(L) * .2}
}
\description{
Simulate data under the twococomo SuSiE model
}
