% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{stick_breaking_N}
\alias{stick_breaking_N}
\title{Number of binomial trials for a stick breaking representation of multinomial}
\usage{
stick_breaking_N(Y)
}
\arguments{
\item{Y}{an N x K matrix of assignment probabilities}
}
\value{
an N x K matrix
}
\description{
Number of binomial trials for a stick breaking representation of multinomial
}
