% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_random_effect_vb.R
\name{sub_re}
\alias{sub_re}
\title{Subtract the predictions from an SER to the random offset}
\usage{
sub_re(o, ser, X)
}
\description{
Subtract the predictions from an SER to the random offset
}
