% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_moococomo_obj.R
\name{t_ind_var.mococomo}
\alias{t_ind_var.mococomo}
\title{Compute individual posterior variance from marginal normal mean model}
\usage{
t_ind_var.mococomo(fit, i)
}
\description{
internal function to compute posterior mean and sds
}
