% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_susie.R
\name{update_b.binsusie}
\alias{update_b.binsusie}
\title{update alpha, mu, and var}
\usage{
update_b.binsusie(
  fit,
  fit_intercept = TRUE,
  fit_prior_variance = TRUE,
  fit_alpha = TRUE,
  update_idx = NULL
)
}
\description{
update alpha, mu, and var
}
