% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise_initialization.R
\name{update_b_hard_select.binsusie}
\alias{update_b_hard_select.binsusie}
\title{update alpha, mu, and var}
\usage{
update_b_hard_select.binsusie(
  fit,
  fit_intercept = TRUE,
  fit_prior_variance = TRUE,
  update_idx = NULL
)
}
\description{
update alpha, mu, and var
}
