% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_ser.R
\name{update_delta.binser}
\alias{update_delta.binser}
\title{update intercept and fixed effect covariates}
\usage{
update_delta.binser(fit, data)
}
\description{
update intercept and fixed effect covariates
}
