% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibss_susie.R
\name{update_model.ibss2m}
\alias{update_model.ibss2m}
\title{update alpha, mu, and var}
\usage{
\method{update_model}{ibss2m}(
  fit,
  data,
  fit_intercept = TRUE,
  fit_prior_variance = TRUE,
  fit_alpha = TRUE,
  fit_xi = TRUE,
  update_idx = NULL,
  track_elbo = TRUE
)
}
\description{
update alpha, mu, and var
}
