% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_distributions.R
\name{update_params.normal}
\alias{update_params.normal}
\title{Update variance of the normal parameter on a grid centered around the current estimate
on a log2 scale}
\usage{
\method{update_params}{normal}(dist, betahat, se, weights)
}
\description{
Update variance of the normal parameter on a grid centered around the current estimate
on a log2 scale
}
