% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_distributions.R
\name{update_params}
\alias{update_params}
\title{Update Params}
\usage{
update_params(x, ...)
}
\arguments{
\item{betahat}{observations}

\item{se}{standard errors of observations}

\item{weights}{weights to weigh each observation by when optimizing}
}
\value{
a new \code{component_distribution} object with update parameters
}
\description{
Generic function for updating the parameters of a \code{component_distribution}
}
