% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_ser.R
\name{update_prior_variance.binser}
\alias{update_prior_variance.binser}
\title{Variational M step for updating prior_variance
Optimize ELBO wrt to prior effect standard deviation}
\usage{
update_prior_variance.binser(fit, k, idx = NULL)
}
\description{
Variational M step for updating prior_variance
Optimize ELBO wrt to prior effect standard deviation
}
