% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_ser.R
\name{update_sigma0.binser}
\alias{update_sigma0.binser}
\title{Variational M step for updating sigma0
Optimize ELBO wrt to prior effect standard deviation}
\usage{
update_sigma0.binser(fit, idx = NULL)
}
\description{
Variational M step for updating sigma0
Optimize ELBO wrt to prior effect standard deviation
}
