% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_cEBMF_obj.R
\name{update_tau.cEBMF}
\alias{update_tau.cEBMF}
\title{Estimate noise value}
\usage{
update_tau.cEBMF(cEBMF.obj)
}
\arguments{
\item{cEBMF}{an CEBMF object}
}
\description{
Estimate noise value
}
