% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_random_effect_vb.R,
%   R/univariate_vb.R
\name{update_tau0}
\alias{update_tau0}
\title{Update prior precision of effect}
\usage{
update_tau0(x, y, o, mu, tau, xi, delta, tau0)

update_tau0(x, y, o, mu, tau, xi, delta, tau0)
}
\description{
Update prior precision of effect

Update prior precision of effect
}
