% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_ser.R
\name{update_var0.binser}
\alias{update_var0.binser}
\title{Variational M step for updating prior_variance
Optimize ELBO wrt to prior effect standard deviation}
\usage{
update_var0.binser(fit)
}
\description{
Variational M step for updating prior_variance
Optimize ELBO wrt to prior effect standard deviation
}
