% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_ser.R
\name{update_xi.binser}
\alias{update_xi.binser}
\title{update for variational parameter parameter xi q(w) = PG(N, xi)}
\usage{
update_xi.binser(fit)
}
\description{
update for variational parameter parameter xi q(w) = PG(N, xi)
}
