% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_susie.R
\name{update_xi.binsusie}
\alias{update_xi.binsusie}
\title{update for variational parameter parameter xi q(w) = PG(N, xi)}
\usage{
update_xi.binsusie(fit)
}
\description{
update for variational parameter parameter xi q(w) = PG(N, xi)
}
