% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_generics.R
\name{update_xi}
\alias{update_xi}
\title{update for variational parameter parameter xi q(w) = PG(N, xi)}
\usage{
update_xi(fit, data)
}
\description{
update for variational parameter parameter xi q(w) = PG(N, xi)
}
