% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvb_ser.R
\name{uvbser}
\alias{uvbser}
\title{UVB SER}
\usage{
uvbser(
  X,
  y,
  o = 0,
  N = rep(1, length(y)),
  Z = NULL,
  scale = FALSE,
  center = TRUE,
  prior_mean = 0,
  prior_variance = 1,
  prior_weights = NULL,
  intercept = TRUE,
  estimate_prior_variance = TRUE,
  check_null_threshold = 0,
  prior_tol = 1e-09,
  prune = FALSE,
  coverage = 0.95,
  min_abs_corr = 0.5,
  max_iter = 100,
  tol = 0.001,
  verbose = FALSE,
  n_purity = 100,
  q = NULL,
  ...
)
}
\arguments{
\item{X}{a n x p matrix of covariates}

\item{y}{an n vector of integer counts, bernoulli/binomial observations}

\item{N}{the number of binomial trials, defaults to 1, may be a scalar or vector of length n}

\item{Z}{fixed effect covaraites (including intercept). If null just a n x 1 matrix of ones}

\item{scale}{if TRUE, scale the columns of X to unit variate}

\item{center}{if TRUE, center the columns of X to mean zero}

\item{prior_mean}{the prior mean of each non-zero element of b. Either a scalar or vector of length L.}

\item{prior_variance}{the prior variance of each non-zero element of b. Either a scalar or vector of length L. If \code{estimate_prior_variance=TRUE} the value provides an initial estimate of prior variances}

\item{prior_weights}{prior probability of selecting each column of X, vector of length p summing to one, or an L x p matrix}

\item{returns}{a fit Binomial SuSiE model, which is compatable with summary functions from \code{susieR} package}
}
\description{
Fit Binomial UVB SER
}
