% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_function.R
\name{compute_data_loglikelihood}
\alias{compute_data_loglikelihood}
\alias{compute_data_loglikelihood.mococomo_normal}
\alias{compute_data_loglikelihood.mococomo_beta}
\title{Compute data likelihood}
\usage{
compute_data_loglikelihood(fit, ...)

\method{compute_data_loglikelihood}{mococomo_normal}(fit, data)

\method{compute_data_loglikelihood}{mococomo_beta}(fit, data)
}
\arguments{
\item{fit}{an MoCoCoMo fit object}
}
\value{
n x K matrix of  log likelihood of each data point for each component
}
\description{
Compute data likelihood
}
\details{
Compute Data Log Likelihood
Compute p(betahat_i| z=k, se_i) i = 1..., n, k=1... K
}
