% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_ser.R
\name{compute_tau}
\alias{compute_tau}
\title{Compute partial posterior variance
use this immediately after updating xi}
\usage{
compute_tau(fit, k, kidx = NULL)
}
\description{
Compute partial posterior variance
use this immediately after updating xi
}
