% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mococomo.R
\name{fit.mococomo}
\alias{fit.mococomo}
\title{Fucntion implementating the mococomo mode}
\usage{
fit.mococomo(data, maxiter = 100, tol = 0.001, max_class, mult = 2)
}
\arguments{
\item{data}{an object of class data_mococomo  see \link{\code{set_data_mococomo}}}

\item{maxiter}{numeric, maximum numerous of itration set to 100 by defaults}

\item{tol}{tolreance in term of chjange in ELBO value for stopping criterion}
}
\description{
Fucntion implementating the mococomo mode
}
\details{
Fucntion implementating the mococomo mode
}
