% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_ser.R
\name{fit_fast_glm}
\alias{fit_fast_glm}
\title{Fit fast GLM}
\usage{
fit_fast_glm(x, y, o, family = "binomial", augment = F)
}
\arguments{
\item{x}{the variable of interest}

\item{y}{the response vector}

\item{o}{fixed offset, length(o) == length(y)}

\item{family}{the family of the glm, e.g. "binomial", "poisson"}
}
\description{
Fit fast GLM, and extract necessary information for SER
}
