% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibss_ser.R
\name{ibss_from_ser}
\alias{ibss_from_ser}
\title{Functions for a generic IBSS algorithm
allows us to specify an SER function to run IBSS with...}
\usage{
ibss_from_ser(
  X,
  y,
  L = 10,
  prior_variance = 1,
  prior_weights = NULL,
  tol = 0.001,
  maxit = 100,
  estimate_intercept = TRUE,
  ser_function = NULL
)
}
\arguments{
\item{ser_function}{a function for fitting SER, takes arguments \code{X}, \code{y}, \code{o} \code{prior_variance}, \code{estimate_intercept}, and \code{prior_weights}}
}
\description{
Functions for a generic IBSS algorithm
allows us to specify an SER function to run IBSS with...
}
