% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cEBMF.R
\name{init_cEBMF}
\alias{init_cEBMF}
\title{Initialize cEBMF object}
\usage{
init_cEBMF(Y, X_l, X_f, K = 1, type_noise = "column_wise")
}
\arguments{
\item{Y}{numerical matrix size NxP}

\item{X_l}{matrix of size NxJ containning covariates affecting the factors}

\item{X_f}{matrix of size PxT containning covariates affecting the factors}

\item{K}{numeric number of factors
@param type_noise specify which kind of noise structure is expected, currently three choices. Whether noise constant accross column ('column_wise'), constant 'constant' or constant across rown 'row_wise'}
}
\value{
a cEBMF object
}
\description{
Initialize cEBMF object
}
