% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_ser.R
\name{map_univariate_regression}
\alias{map_univariate_regression}
\title{Map univariate regression}
\usage{
map_univariate_regression(
  X,
  y,
  off = NULL,
  estimate_intercept = T,
  family = "binomial",
  augment = F
)
}
\arguments{
\item{X}{n x p design matrix}

\item{y}{n vector response}

\item{estimate_intercept}{boolean to fit intercept}

\item{family}{family for glm, e.g. binomial}

\item{o}{optional offset, vector of length n}
}
\value{
list containing intercept and effect estimates,
standard errors, and likelihood ratios
}
\description{
Fit univariate glm for each column of X
}
