% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ser_from_univariate.R
\name{ser_from_univariate}
\alias{ser_from_univariate}
\title{Take a function for fitting MLE of univariate regression and make an SER function}
\usage{
ser_from_univariate(uni_fun)
}
\arguments{
\item{uni_fun}{a function for fitting univariate regression
must take arguments: x, y, o, estimate_intercept
additional arguments passed via ...
must return: a list with items lbf = log bayes factor, mu = posterior mean, var = posterior variance, intercept=intercept
note that posterior variance doesn't figure in computations so you can set it to a dummy value (e.g. 0)}
}
\description{
this is a utlity function for generating SER fit function compatible with \code{ibss_from_ser}
if you have a function for fitting a univariate regression this will create a function for fitting an SER
it relied on an asymptotic approximation-- so if you have betahat, standard error, and likelihood ratio you can use this
}
