% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_data_mococomo}
\alias{set_data_mococomo}
\title{Prepraring data for mococomo fit}
\usage{
set_data_mococomo(beta, se, X, ...)
}
\arguments{
\item{betahat}{numeric, vecotr of regression coefficients  list of containing the following element see}

\item{maxiter}{numeric, maximum numerous of itration set to 100 by defaults}

\item{tol}{tolreance in term of chjange in ELBO value for stopping criterion}
}
\description{
Prepraring data for mococomo fit
}
\details{
Prepraring data for mococomo fit
}
